Delete extended trace and move the stack trace printer to logging
